+++ zarrs_chunk_grid/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Zero sized dimensions are no longer considered _unlimited_
+  - *Behavioural Change*: `array_indices_inbounds` and `chunk_indices_inbounds` now always return `false` for zero-sized arrays
+
 ## [0.5.1] - 2026-02-08
 
 ### Added